Refactors testutil into separate files
